feat 36761802: added folder utils for helpers
