+++ packages/pkg.module.whiteboard/utils/customConfig.ts
