+++ packages/pkg.module.whiteboard/utils/imageStore.ts
